Rename folders: `src` to `extension` and `app to `src`
