+++ extension/tsconfig.json
@@ -1,0 +1,10 @@
+{
+  "compilerOptions": {
+    "module": "commonjs",
+    "target": "es2020",
+    "lib": ["es2020"],
+    "outDir": "../out",
+    "sourceMap": true,
+    "strict": true
+  }
+}
