+++ src/Viewer.tsx
@@ -2,7 +2,7 @@ import { App } from '@h5web/app';
 import { H5WasmBufferProvider } from '@h5web/h5wasm';
 import { suspend } from 'suspend-react';
 import { getExportURL, getPlugin } from './utils';
-import { type FileInfo } from '../src/models.js';
+import { type FileInfo } from '../extension/models.js';
 
 interface Props {
   fileInfo: FileInfo;
