+++ src/vscode-api.ts
