+++ vite.config.js
@@ -1,9 +1,9 @@
 import { defineConfig } from 'vite';
-import react from '@vitejs/plugin-react';
+import react from '@vitejs/plugin-react-swc';
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  root: './app',
+  root: './src',
   publicDir: false,
   plugins: [react()],
   build: {
